split and rename permission tests files so they can be found easier
